TISJEW-4488:Removing Model Refresh From List of Objects:883846

+++ ext/marketplacecommerceservices/src/com/tisl/mpl/marketplacecommerceservices/service/impl/ExchangeGuideServiceImpl.java
@@ -40,6 +40,7 @@ public class ExchangeGuideServiceImpl implements ExchangeGuideService
 	//SONAR FIX JEWELLERY
 	protected static final Logger LOG = Logger.getLogger(ExchangeGuideServiceImpl.class);
 	private static final String ERROR_OCCURED = "Error Occured At ";
+	private static final String WHILE_SAVING = "While Saving";
 
 	@Autowired
 	private PersistentKeyGenerator temporaryExchangeId;
@@ -176,7 +177,7 @@ public class ExchangeGuideServiceImpl implements ExchangeGuideService
 		}
 		catch (final ModelSavingException e)
 		{
-			LOG.error(ERROR_OCCURED + "While Saving" + "getExchangeID");
+			LOG.error(ERROR_OCCURED + WHILE_SAVING + "getExchangeID");
 		}
 		catch (final EtailNonBusinessExceptions e)
 		{
@@ -242,13 +243,11 @@ public class ExchangeGuideServiceImpl implements ExchangeGuideService
 				exListToSave.add(ex);
 			}
 			modelService.saveAll(exListToSave);
-			//Added for EQA Comments
-			modelService.refresh(exListToSave);
 			isSaved = true;
 		}
 		catch (final ModelSavingException e)
 		{
-			LOG.error(ERROR_OCCURED + "While Saving" + "changePincode");
+			LOG.error(ERROR_OCCURED + WHILE_SAVING + "changePincode");
 		}
 		catch (final EtailNonBusinessExceptions e)
 		{
@@ -298,7 +297,7 @@ public class ExchangeGuideServiceImpl implements ExchangeGuideService
 		}
 		catch (final ModelSavingException e)
 		{
-			LOG.error(ERROR_OCCURED + "While Saving" + "removeFromTransactionTable");
+			LOG.error(ERROR_OCCURED + WHILE_SAVING + "removeFromTransactionTable");
 		}
 
 		catch (final EtailNonBusinessExceptions e)
@@ -454,24 +453,19 @@ public class ExchangeGuideServiceImpl implements ExchangeGuideService
 
 
 			modelService.saveAll(exModList);
-			//Added for EQA Comments
-			modelService.refresh(exModList);
+
 			modelService.saveAll(childOrderEntryList);
-			//Added for EQA Comments
-			modelService.refresh(childOrderEntryList);
+
 			modelService.saveAll(childModfList);
-			//Added for EQA Comments
-			modelService.refresh(childModfList);
+
 			modelService.saveAll(parentEntryList);
-			//Added for EQA Comments
-			modelService.refresh(parentEntryList);
+
 			modelService.removeAll(exTraxRemovList);
-			//Added for EQA Comments
-			modelService.refresh(exTraxRemovList);
+
 		}
 		catch (final ModelSavingException e)
 		{
-			LOG.error(ERROR_OCCURED + "While Saving" + "getExchangeRequestID2");
+			LOG.error(ERROR_OCCURED + WHILE_SAVING + "getExchangeRequestID2");
 		}
 		catch (final EtailNonBusinessExceptions e)
 		{
@@ -560,17 +554,18 @@ public class ExchangeGuideServiceImpl implements ExchangeGuideService
 				}
 			}
 
-			modelService.saveAll(exModList);
-			//Added for EQA Comments
-			modelService.refresh(exModList);
-
-			modelService.removeAll(exTraxRemovList);
-			//Added for EQA Comments
-			modelService.refresh(exTraxRemovList);
+			if (CollectionUtils.isNotEmpty(exModList))
+			{
+				modelService.saveAll(exModList);
+			}
+			if (CollectionUtils.isNotEmpty(exTraxRemovList))
+			{
+				modelService.removeAll(exTraxRemovList);
+			}
 		}
 		catch (final ModelSavingException e)
 		{
-			LOG.error(ERROR_OCCURED + "While Saving" + "getExchangeRequestID");
+			LOG.error(ERROR_OCCURED + WHILE_SAVING + "getExchangeRequestID");
 		}
 		catch (final EtailNonBusinessExceptions e)
 		{
@@ -603,15 +598,17 @@ public class ExchangeGuideServiceImpl implements ExchangeGuideService
 				}
 
 			}
-			modelService.saveAll(entryUpdate);
-			//Added for EQA Comments
-			modelService.refresh(entryUpdate);
+			if (CollectionUtils.isNotEmpty(entryUpdate))
+			{
+				modelService.saveAll(entryUpdate);
+			}
+
 
 			removeFromTransactionTable(removeExchangeIdList, null, cartModel);
 		}
 		catch (final ModelSavingException e)
 		{
-			LOG.error(ERROR_OCCURED + "While Saving" + "removeExchangefromCart");
+			LOG.error(ERROR_OCCURED + WHILE_SAVING + "removeExchangefromCart");
 		}
 		catch (final EtailNonBusinessExceptions e)
 		{
@@ -650,14 +647,17 @@ public class ExchangeGuideServiceImpl implements ExchangeGuideService
 				listToSave.add(exTrax);
 			}
 
-			modelService.saveAll(listToSave);
-			//Added for EQA Comments
-			modelService.refresh(listToSave);
+			if (CollectionUtils.isNotEmpty(listToSave))
+			{
+				modelService.saveAll(listToSave);
+
+			}
+
 
 		}
 		catch (final ModelSavingException e)
 		{
-			LOG.error(ERROR_OCCURED + "While Saving" + "changeGuidforCartMerge");
+			LOG.error(ERROR_OCCURED + WHILE_SAVING + "changeGuidforCartMerge");
 		}
 		catch (final EtailNonBusinessExceptions e)
 		{
